requirements.txt
asgiref==3.8.1
black==24.8.0
cfgv==3.4.0
click==8.1.7
colorama==0.4.6
distlib==0.3.8
Django==4.2.15
django-filter==2.4.0
djangorestframework==3.15.*
filelock==3.15.4
identify==2.6.0
Markdown==3.6
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pre-commit==3.8.0
pytz==2024.1
PyYAML==6.0.2
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
virtualenv==20.26.3
